[ERROR!] Aggiunto Utilities al vertxServer e vertxClient. Refactoring rooms...
