+++ core/src/main/scala/it/cwmp/utils/Loggable.scala
@@ -1,0 +1,8 @@
+package it.cwmp.utils
+
+import com.typesafe.scalalogging.Logger
+
+trait Loggable {
+
+  val logger: Logger = Logger(this.getClass.getName)
+}
