+++ server/rooms/src/main/scala/it/cwmp/RoomsServiceMain.scala
@@ -4,7 +4,7 @@ import com.typesafe.scalalogging.Logger
 import io.vertx.lang.scala.VertxExecutionContext
 import io.vertx.scala.core.Vertx
 import it.cwmp.controller.client.RoomReceiverApiWrapper
-import it.cwmp.controller.rooms.RoomsServiceVerticle
+import it.cwmp.services.rooms.RoomsServiceVerticle
 import it.cwmp.services.wrapper.AuthenticationApiWrapper
 
 import scala.util.{Failure, Success}
